Moved parse_bounds into source
